% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{cytoprofiling_to_seurat}
\alias{cytoprofiling_to_seurat}
\title{Export cytoprofiling cell table to Seurat object}
\usage{
cytoprofiling_to_seurat(
  df,
  panel_json,
  use_gene_names = FALSE,
  split_assays = FALSE
)
}
\arguments{
\item{df}{Cell table (see load_cytoprofiling)}

\item{panel_json}{JSON data for Panel.json (see load_panel)}

\item{use_gene_names}{If TRUE, use gene names as rownames, otherwise use target names. Duplicate gene names will be removed.}

\item{split_assays}{If TRUE, split the transcript and protein barcoding targets into separate Seurat assays ("RNA" and "Protein", respectively).}
}
\value{
Seurat object
}
\description{
Export cytoprofiling cell table to Seurat object
}
