% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_cells}
\alias{filter_cells}
\title{Perform default filtering of cells (filter by area, assigned counts, and assigned rate)}
\usage{
filter_cells(df, batch_names = c(), well_names = c())
}
\arguments{
\item{df}{Cell table}

\item{batch_names}{List of batch names}

\item{well_names}{List of well names}
}
\value{
Filtered cell table
}
\description{
Perform default filtering of cells (filter by area, assigned counts, and assigned rate)
}
