% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_cells_by_area_percentile}
\alias{filter_cells_by_area_percentile}
\title{Filter cells to remove outliers by area}
\usage{
filter_cells_by_area_percentile(
  df,
  well_names = c(),
  min_threshold = 3,
  max_threshold = 97
)
}
\arguments{
\item{df}{Cell table}

\item{well_names}{List of well names}

\item{min_threshold}{Lower percentile threshold}

\item{max_threshold}{Upper percentile threshold}
}
\value{
Filtered cell table
}
\description{
Filter cells to remove outliers by area
}
