% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_cells_by_assigned_rate}
\alias{filter_cells_by_assigned_rate}
\title{Filter cells on assigned rate}
\usage{
filter_cells_by_assigned_rate(
  df,
  batch_names = c(),
  well_names = c(),
  assigned_rate_threshold = 0.5
)
}
\arguments{
\item{df}{Cell table}

\item{batch_names}{List of batch names}

\item{well_names}{List of well names}

\item{assigned_rate_threshold}{Assigned rate threshold}
}
\value{
Filtered cell table
}
\description{
Cells must exceed the assignment rate threshold in all batches
to be kept
}
