% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_targets}
\alias{get_all_targets}
\title{Get all targets for a batch}
\usage{
get_all_targets(df, batch)
}
\arguments{
\item{df}{Cell table}

\item{batch}{Batch name}
}
\value{
List of all targets
}
\description{
Get all targets for a batch
}
