% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_barcoding_batches}
\alias{get_barcoding_batches}
\title{Get a list of barcoding batches}
\usage{
get_barcoding_batches(df)
}
\arguments{
\item{df}{Cell table}
}
\value{
List of barcoding batches
}
\description{
Get a list of barcoding batches present in the cell table
}
