% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_normalization_targets}
\alias{get_default_normalization_targets}
\title{Get normalization targets for a batch}
\usage{
get_default_normalization_targets(df, batch)
}
\arguments{
\item{df}{Cell table}

\item{batch}{Batch name}
}
\value{
List of normalization targets
}
\description{
Get a standard list of targets to use for normalization in a batch
}
