% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_cytoprofiling}
\alias{load_cytoprofiling}
\title{Load a parquet file into a data frame}
\usage{
load_cytoprofiling(input_filename)
}
\arguments{
\item{input_filename}{Path to the parquet file}
}
\value{
Data frame
}
\description{
Calls arrow::read_parquet to load cytoprofiling
parquet file into a dataframe
}
