% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_panel}
\alias{load_panel}
\title{Load a Panel.json file}
\usage{
load_panel(input_filename)
}
\arguments{
\item{input_filename}{Path to the Panel.json file}
}
\value{
JSON data as a list
}
\description{
Calls rjson::fromJSON to load a Panel.json file
}
