% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_cells_by_aggregated_counts}
\alias{normalize_cells_by_aggregated_counts}
\title{Normalized cell table by aggregated counts}
\usage{
normalize_cells_by_aggregated_counts(
  df,
  batch_names = c(),
  normalization_targets = c()
)
}
\arguments{
\item{df}{Cell table}

\item{batch_names}{List of batch names}

\item{normalization_targets}{List of normalization targets}
}
\value{
Normalized cell table
}
\description{
Normalize each cell by number of counts in each batch. If list of normalization targets is
provided, normalize by the sum of counts in the targets. Otherwise, normalize by the sum of
default normalization targets (see get_default_normalization_targets).
}
