% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_cytoprofiling}
\alias{normalize_cytoprofiling}
\title{Perform default normalization for cells and samples}
\usage{
normalize_cytoprofiling(df, batch_names = c(), well_names = c())
}
\arguments{
\item{df}{Cell table}

\item{batch_names}{List of batch names}

\item{well_names}{List of well names}
}
\value{
Normalized cell table
}
\description{
Normalize cells by assigned counts per batch and normalize samples by median of ratios.
}
