% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{cytoprofiling_to_seurat}
\alias{cytoprofiling_to_seurat}
\title{Export cytoprofiling cell table to Seurat object}
\usage{
cytoprofiling_to_seurat(df)
}
\arguments{
\item{df}{Cell table}
}
\value{
Seurat object
}
\description{
Export cytoprofiling cell table to Seurat object
}
